Filter and decode the TPS1 data
This code with filter and decode the TPS1 data sent from bmm ecu. It takes the hex data and puts it in right order (originally sent in little endian format), then will decode from hex to decimal and finally multiply by the factor of .01 to get the true tps readout. This is just a test code so you don't have to be running the car to see if your canBus is working correctly,