.dbc file for BMM ecu
This is the .dbc (database container) file for the Beer Money Motorsports ecu. It contains all of the info for messages and signals that can be trasmitted over the can network. For an easier to read version of this with the name, CAN ID, dlc, start, legnth, factor, offset, min/max, and unit of measurement. This can be found at https://wiki.fome.tech/Advanced-Features/CAN/